+++ R/check_args.R
@@ -1,14 +1,15 @@
-# Copyright (c) Cancer Council NSW, 2018-2023. All rights reserved.
+# SPDX-FileCopyrightText: 2023 Stephen Wade <stephematician@gmail.com>
+# SPDX-License-Identifier: MIT
 
-# Helpers to check arguments in call to smirf()
+# Helpers to check arguments in call to mimputest()
 
-check_args <- function(data, model, n, verbose, fn_init,
-                       stop_measure, loop_limit, overrides, clean_step) {
+check_args <- function(data, model, n, fn_init, stop_measure, loop_limit,
+                       overrides, clean_step, keep_forests, verbose) {
 
     check_arg_types(
-        data=data, model=model, n=n, verbose=verbose,
-        fn_init=fn_init, stop_measure=stop_measure, loop_limit=loop_limit,
-        overrides=overrides, clean_step=clean_step
+        data=data, model=model, n=n, fn_init=fn_init, stop_measure=stop_measure,
+        loop_limit=loop_limit, overrides=overrides, clean_step=clean_step,
+        keep_forests=keep_forests, verbose=verbose
     )
 
     msgs <- is_valid_data(data=data)
@@ -18,28 +19,35 @@ check_args <- function(data, model, n, verbose, fn_init,
 
     msgs %<>% c(
         msg_if_not(length(n) == 1, '\'n\' is not a scalar value'),
-        msg_if_not(n > 0, '\'n\' must be non-negative'),
-        msg_if_not(length(verbose) == 1), msg_if_not(length(loop_limit) == 1),
-        msg_if_not(loop_limit >= 0)
+        msg_if_not(n > 0, '\'n\' must be non-negative')
     )
 
     msgs %<>%
         is_valid_fn_init(fn_init=fn_init, msgs=.) %>%
-        is_valid_stop_measure(stop_measure=stop_measure, msgs=.) %>%
-        is_valid_clean_step(data=data, clean_step=clean_step, msgs=.)
+        is_valid_stop_measure(stop_measure=stop_measure, msgs=.)
+
+    msgs %<>% c(
+        msg_if_not(loop_limit >= 0), msg_if_not(length(keep_forests) == 1)
+    )
+
+    msgs %<>% is_valid_clean_step(data=data, clean_step=clean_step, msgs=.)
+
+    msgs %<>% c(
+        msg_if_not(length(verbose) == 1), msg_if_not(length(loop_limit) == 1)
+    )
 
     if (length(msgs) > 0)
-        stop('The following pre-conditions failed: ',
-             paste(msgs, collapse='; '), '.')
+        stop('The following pre-conditions failed:\n ',
+             paste('-   ', msgs[-length(msgs)], collapse=';\n '),
+             '-   ', msgs[length(msgs)], '.')
 
     invisible()
 
 }
 
 
-check_arg_types <- function(data, model, n, verbose, fn_init,
-                            stop_measure, loop_limit, overrides,
-                            clean_step) {
+check_arg_types <- function(data, model, n, fn_init, stop_measure, loop_limit,
+                            overrides, clean_step, keep_forests, verbose) {
 
     msgs <- c(msg_if_not(is.data.frame(data)), msg_if_not(is.numeric(n)))
 
@@ -47,15 +55,15 @@ check_arg_types <- function(data, model, n, verbose, fn_init,
         msgs %<>% c(msg_if_not(is.logical(model)), msg_if_not(is.matrix(model)))
 
     msgs %<>% c(
-        msg_if_not(is.logical(verbose)),
         msg_if_not(is.function(fn_init)), msg_if_not(is.function(stop_measure)),
         msg_if_not(is.numeric(loop_limit)), msg_if_not(is.list(overrides)),
-        msg_if_not(is.list(clean_step))
+        msg_if_not(is.list(clean_step)), msg_if_not(is.logical(keep_forests)),
+        msg_if_not(is.logical(verbose))
     )
 
     if (length(msgs > 0))
         stop('The following type requirements failed: ',
-             paste(type_check_msg, collapse='; '), '.')
+             paste(msgs, collapse='; '), '.')
 
     invisible()