+++ R/post_process.R
@@ -1,77 +1,25 @@
-# Post process a single imputation
-post_process_missforest <- function(res, to_categories) {
+# SPDX-FileCopyrightText: 2023 Stephen Wade <stephematician@gmail.com>
+# SPDX-License-Identifier: MIT
 
-    iterations_ <- unname(mapply(list,
-                                 iteration=seq_len(1L + res$iterations) - 1L,
-                                 SIMPLIFY=F))
+# post-process the imputed values of a whole chain
+post_process <- function(result) {
 
-    aggregates <- mapply(post_process_iteration,
-                         imputed=res$imputed,
-                         iteration=iterations_,
-                         MoreArgs=list(to_categories=to_categories),
-                         SIMPLIFY=F)
+    each_iteration <- mapply(post_process_iteration,
+                             imputed=result$imputed,
+                             iteration=seq_len(1L + result$iterations) - 1L)
 
-    # bind all the iterations together
-    c(res,
-      list(frequency=do.call(rbind,
-                             lapply(aggregates, getElement, 'frequency')),
-           statistics=do.call(rbind,
-                              lapply(aggregates, getElement, 'statistics'))))
+    apply(each_iteration, 1, do.call, what=rbind, simplify=F)
 
 }
 
-# Post process a single iteration
-post_process_iteration <- function(imputed,
-                                   to_categories=list(),
-                                   iteration=data.frame(iteration=0)) {
+post_process_iteration <- function(imputed, iteration) {
 
-    categorical_data <- names(imputed)[!sapply(imputed, is.numeric) |
-                                           !!sapply(imputed, is.integer)]
-
-    list(frequency=data.frame(iteration,
-                              frequency_of_imputed(imputed[categorical_data],
-                                                   to_categories)),
-         statistics=data.frame(iteration,
-                               statistics_of_imputed(imputed)))
+    list(statistics=statistics_of_imputed(imputed=imputed, iteration=iteration),
+         frequency=frequency_of_imputed(imputed=imputed, iteration=iteration))
 
 }
-#' Calculate statistics of imputed data
-#'
-#' Calculates selected statistics (measures) of imputed data
-#'
-#' Calculates statistics or measures for each imputed variable according to its
-#' type, which is one of categorical, ordinal or continuous:
-#' \describe{
-#'   \item{categorical}{
-#'     calculate the \code{\link{entropy}} of non-numeric or integer data;
-#'   }
-#'   \item{ordinal}{
-#'     calculate the dispersion using \code{\link{leiks_D}} of ordered or
-#'     integer data;
-#'   }
-#'   \item{continuous}{
-#'     calculate the mean and variance of numeric data;
-#'   }
-#' }
-#'
-#' These statistics are intended to be used to monitor convergence of the
-#' missForest or MICE procedure.
-#'
-#' @param imputed named list; each item in the last contains the imputed values
-#'   of the named item.
-#' @return data.frame; a data.frame with three columns;
-#' \describe{
-#'   \item{`variable`}{name of the imputed data as a factor;}
-#'   \item{`value`}{value of the measure/statistic;}
-#'   \item{`measure`}{the name of the measure as a factor, e.g. 'entropy'.}
-#' }
-#'
-#' @seealso entropy leiks_D
-#'
-#' @keywords internal
-#' @importFrom rlang sym
-#' @md
-statistics_of_imputed <- function(imputed) {
+
+statistics_of_imputed <- function(imputed, iteration) {
 
     vars <- names(imputed)
 
@@ -79,8 +27,8 @@ statistics_of_imputed <- function(imputed) {
         rbind,
         mapply(data.frame,
                variable=each_var,
-               value=lapply(imputed[each_var], eval(rlang::sym(aggregator))),
-               MoreArgs=list(measure=aggregator),
+               value=lapply(imputed[each_var], match.fun(aggregator)),
+               MoreArgs=list(measure=aggregator, iteration=iteration),
                SIMPLIFY=F, USE.NAMES=F)
     )
 
@@ -98,45 +46,24 @@ statistics_of_imputed <- function(imputed) {
 
 }
 
+#' @importFrom stats setNames
+#' @noRd
+frequency_of_imputed <- function(imputed, iteration=NA_integer_) {
 
-#' Frequency of imputed values of categorical data
-#'
-#' Frequency of imputed values of categorical data
-#'
-#' @param imputed named list;
-#' @param to_categories named list;
-#'
-#' @section To-do:
-#' \itemize{
-#'   \item fully document \code{frequency_of_imputed}
-#'   \item write tests for \code{frequency_of_imputed}
-#' }
-#'
-#' @keywords internal
-frequency_of_imputed <- function(imputed, to_categories=list()) {
+    vars <- names(imputed)
 
-    variable_names_ <- unname(mapply(list,
-                                     variable=names(imputed),
-                                     SIMPLIFY=F))
+    not_numeric_var <- !sapply(imputed, is.numeric)
+    not_integer_var <- !sapply(imputed, is.integer)
 
-    categories <- lapply(imputed, levels)
-    categories[names(to_categories)] <- lapply(to_categories, names)
-
-    no_categories <- names(categories)[sapply(categories, is.null)]
-    if (length(no_categories) > 0)
-        stop(paste0('no category labels found for\n  - ',
-                    paste0(no_categories, collapse=', '), '.'))
-
-    do.call(rbind,
-            # make data frame for each variable
-            mapply(function(name_, x)
-                       data.frame(name_,
-                                  setNames(data.frame(table(x)),
-                                           nm=c('category',
-                                                'frequency'))),
-                   variable_names_,
-                   mapply(factor, imputed, levels=categories, SIMPLIFY=F),
-                   SIMPLIFY=F))
+    cat_vars <- vars[not_numeric_var | !not_integer_var]
+    mapply(data.frame,
+           variable=cat_vars,
+           lapply(imputed[cat_vars], table) %>%
+               lapply(data.frame) %>%
+               lapply(setNames, nm=c('category', 'frequency')),
+           MoreArgs=list(iteration=iteration),
+           SIMPLIFY=FALSE, USE.NAMES=FALSE) %>%
+    do.call(what=rbind)
 
 }