+++ tests/testthat/test-most_frequent_values.R
@@ -1,4 +1,5 @@
-# Copyright (c) Cancer Council NSW, 2018-2023. All rights reserved.
+# SPDX-FileCopyrightText: 2023 Stephen Wade <stephematician@gmail.com>
+# SPDX-License-Identifier: MIT
 
 test_that('class is preserved', {
     data <- data.frame(x=numeric(1), y=integer(1), z=logical(1),