+++ tests/testthat/test-row_sample_n.R
@@ -1,3 +1,5 @@
+# SPDX-FileCopyrightText: 2023 Stephen Wade <stephematician@gmail.com>
+# SPDX-License-Identifier: MIT
 
 test_that('single column', {